refactor protocols and add test
