+++ test/unit/protocols/grpc/GRPCClient.test.ts
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { GRPCClient } from "../../../src";
+import { GRPCClient } from "../../../../src";
 
 describe("grpc", () => {
   it("getEndpoint should remove http and https from endpoint", () => {
