+++ test/unit/protocols/grpc/binding.test.ts
@@ -11,8 +11,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import GRPCClientBinding from "../../../src/implementation/Client/GRPCClient/binding";
-import { InvokeBindingRequest } from "../../../src/proto/dapr/proto/runtime/v1/dapr_pb";
+import GRPCClientBinding from "../../../../src/implementation/Client/GRPCClient/binding";
+import { InvokeBindingRequest } from "../../../../src/proto/dapr/proto/runtime/v1/dapr_pb";
 
 describe("grpc/binding", () => {
   describe("send should call invokeBinding with correct arguments", () => {
