+++ test/unit/protocols/grpc/pubsub.test.ts
@@ -11,8 +11,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import GRPCClientPubSub from "../../../src/implementation/Client/GRPCClient/pubsub";
-import { PublishEventRequest } from "../../../src/proto/dapr/proto/runtime/v1/dapr_pb";
+import GRPCClientPubSub from "../../../../src/implementation/Client/GRPCClient/pubsub";
+import { PublishEventRequest } from "../../../../src/proto/dapr/proto/runtime/v1/dapr_pb";
 
 describe("grpc/pubsub", () => {
   describe("publish should call publishEvent with correct arguments", () => {
